requirements.txt
requests
python-dotenv
IPython
